refactor gameConsole 패키지 분리
